pyproject.toml
[project]
name = "prog-seg-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
